pacotão da madrugada

+++ VendasApi/Vendas.WebApi/Controllers/CabecalhoVendasController.cs
@@ -1,13 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Vendas.WebApi.Controllers;
+﻿using Microsoft.AspNetCore.Mvc;
 using Vendas.Infrastructure.Entity;
 using Microsoft.AspNetCore.Authorization;
 using Vendas.Infrastructure.Repository.Interfaces;
+using Vendas.Infrastructure.Entities.Enums;
 
 namespace Vendas.WebApi.Controllers
 {
@@ -26,14 +21,23 @@ namespace Vendas.WebApi.Controllers
         [HttpPost]
         public IActionResult Add([FromBody]CabecalhoVenda entity)
         {
-            return Ok(_cabecalhoVendaRepository.Save(entity));
+            if (ModelState.IsValid)
+                return Ok(_cabecalhoVendaRepository.Save(entity));
+            else
+                return BadRequest(new { error = "Request inválido" });
         }
 
+        [Authorize("Bearer", Roles = "Admin")]
         [HttpDelete("{id}")]
         public IActionResult Delete(string id)
         {
-            _cabecalhoVendaRepository.Delete(id);
-            return Ok(new { data = "Deletado" });
+            if (ModelState.IsValid)
+            {
+                _cabecalhoVendaRepository.Delete(id);
+                return Ok(new { data = "Deletado" });
+            }
+            else
+                return BadRequest(new { error = "Request inválido" });
         }
 
         [HttpGet]
@@ -45,13 +49,55 @@ namespace Vendas.WebApi.Controllers
         [HttpGet("{id}")]
         public IActionResult FindById(string id)
         {
-            return Ok(_cabecalhoVendaRepository.FindById(id));
+            if (ModelState.IsValid)
+                return Ok(_cabecalhoVendaRepository.FindById(id));
+            else
+                return BadRequest(new { error = "Request inválido" });
+        }
+
+        [HttpGet("usuario/{id}")]
+        public IActionResult FindByIdUsuario(string id)
+        {
+            if (ModelState.IsValid)
+                return Ok(_cabecalhoVendaRepository.FindByIdUsuario(id));
+            else
+                return BadRequest(new { error = "Request inválido" });
+        }
+
+        [HttpGet("usuario/{id}/pendente")]
+        public IActionResult FindByIdUsuarioAndPendente(string id)
+        {
+            if (ModelState.IsValid)
+                return Ok(_cabecalhoVendaRepository.FindByIdUsuarioAndStatus(id, Situacao.Pendente));
+            else
+                return BadRequest(new { error = "Request inválido" });
+        }
+
+        [HttpGet("usuario/{id}/aprovado")]
+        public IActionResult FindByIdUsuarioAndAprovado(string id)
+        {
+            if (ModelState.IsValid)
+                return Ok(_cabecalhoVendaRepository.FindByIdUsuarioAndStatus(id, Situacao.Aprovado));
+            else
+                return BadRequest(new { error = "Request inválido" });
+        }
+
+        [HttpGet("usuario/{id}/reprovado")]
+        public IActionResult FindByIdUsuarioAndReprovado(string id)
+        {
+            if (ModelState.IsValid)
+                return Ok(_cabecalhoVendaRepository.FindByIdUsuarioAndStatus(id, Situacao.Reprovado));
+            else
+                return BadRequest(new { error = "Request inválido" });
         }
 
         [HttpPut]
         public IActionResult Update([FromBody]CabecalhoVenda entity)
         {
-            return Ok(_cabecalhoVendaRepository.Update(entity));
+            if (ModelState.IsValid)
+                return Ok(_cabecalhoVendaRepository.Update(entity));
+            else
+                return BadRequest(new { error = "Request inválido" });
         }
     }
 }
